requirements.txt
python-telegram-bot==20.7
Pillow==10.4.0
requests==2.32.3
aiohttp==3.10.11
APScheduler==3.10.4
